pyproject.toml
[project]
name = "multi-lingual-mcq-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nltk>=3.9.1",
    "indic-nlp-library>=0.92",
    "langdetect>=1.0.9",
    "nltk>=3.9.1",
    "pip>=25.0.1",
    "spacy>=3.8.4",
    "stanza>=1.10.1",
    "torch>=2.6.0",
    "transformers>=4.50.1",
    "flair==0.15.1",
    "gensim==4.3.2",
    "scipy==1.12.0",
    "sentence-transformers>=4.0.2",
    "rouge>=1.0.1",
    "tqdm>=4.67.1",
    "sentencepiece>=0.2.0",
    "scikit-learn>=1.6.1",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "numpy>=1.26.4",
    "fasttext>=0.9.3",
    "mwparserfromhell>=0.6.6",
    "datasets>=3.5.0",
    "peft>=0.15.2",
    "tensorflow>=2.18.1",
    "tf-keras>=2.18.0",
    "googletrans>=4.0.2",
    "anvil-uplink>=0.5.2",
    "seaborn>=0.13.2",
    "keybert>=0.9.0",
    "rouge-score>=0.1.2",
]
